requirements.txt
streamlit
requests
python-dotenv
Pillow
gnews4py
